Add shared module
