+++ Cargo.toml
@@ -30,6 +30,7 @@ members = [
     "unix_integration/resolver_kanidm",
     "rlm_kanidm/entrypoint",
     "rlm_kanidm/module",
+    "rlm_kanidm/shared",
     "server/daemon",
     "server/lib",
     "server/lib-macros",
@@ -140,6 +141,7 @@ kanidm_proto = { path = "./proto", version = "=1.11.0-dev" }
 kanidm_utils_users = { path = "./libs/users", version = "=1.11.0-dev" }
 nss_sparkle_common = { path = "./unix_integration/nss_sparkle_common", version = "=1.11.0-dev" }
 pam_sparkle_common = { path = "./unix_integration/pam_sparkle_common", version = "=1.11.0-dev" }
+rlm_kanidm_shared = { path = "./rlm_kanidm/shared", version = "=1.11.0-dev" }
 sparkle_unix_common = { path = "./unix_integration/common", version = "=1.11.0-dev" }
 sparkle_resolver_common = { path = "./unix_integration/resolver_common", version = "=1.11.0-dev" }
 scim_proto = { path = "./libs/scim_proto", version = "=1.11.0-dev" }
