+++ rlm_kanidm/shared/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "rlm_kanidm_shared"
+description = "Kanidm FreeRADIUS module and entrypoint shared library"
+
+version = { workspace = true }
+authors = { workspace = true }
+rust-version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+
+[lib]
+name = "rlm_kanidm_shared"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+
