+++ rlm_kanidm/shared/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod config;
+
